remove unnecessary boxing and unboxing

+++ src/main/java/net/minecraft/block/BlockRedstoneWire.java
@@ -31,7 +31,7 @@ public class BlockRedstoneWire extends Block {
 
     public BlockRedstoneWire() {
         super(Material.circuits);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(NORTH, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(EAST, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(SOUTH, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(WEST, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(POWER, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(NORTH, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(EAST, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(SOUTH, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(WEST, BlockRedstoneWire.EnumAttachPosition.NONE).withProperty(POWER, 0));
         this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 0.0625F, 1.0F);
     }
 
@@ -90,7 +90,7 @@ public class BlockRedstoneWire extends Block {
 
     public int colorMultiplier(IBlockAccess worldIn, BlockPos pos, int renderPass) {
         IBlockState iblockstate = worldIn.getBlockState(pos);
-        return iblockstate.getBlock() != this ? super.colorMultiplier(worldIn, pos, renderPass) : this.colorMultiplier(iblockstate.getValue(POWER).intValue());
+        return iblockstate.getBlock() != this ? super.colorMultiplier(worldIn, pos, renderPass) : this.colorMultiplier(iblockstate.getValue(POWER));
     }
 
     public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
@@ -111,7 +111,7 @@ public class BlockRedstoneWire extends Block {
 
     private IBlockState calculateCurrentChanges(World worldIn, BlockPos pos1, BlockPos pos2, IBlockState state) {
         IBlockState iblockstate = state;
-        int i = state.getValue(POWER).intValue();
+        int i = state.getValue(POWER);
         int j = 0;
         j = this.getMaxCurrentStrength(worldIn, pos2, j);
         this.canProvidePower = false;
@@ -154,7 +154,7 @@ public class BlockRedstoneWire extends Block {
         }
 
         if (i != j) {
-            state = state.withProperty(POWER, Integer.valueOf(j));
+            state = state.withProperty(POWER, j);
 
             if (worldIn.getBlockState(pos1) == iblockstate) {
                 worldIn.setBlockState(pos1, state, 2);
@@ -234,7 +234,7 @@ public class BlockRedstoneWire extends Block {
         if (worldIn.getBlockState(pos).getBlock() != this) {
             return strength;
         } else {
-            int i = worldIn.getBlockState(pos).getValue(POWER).intValue();
+            int i = worldIn.getBlockState(pos).getValue(POWER);
             return i > strength ? i : strength;
         }
     }
@@ -262,7 +262,7 @@ public class BlockRedstoneWire extends Block {
         if (!this.canProvidePower) {
             return 0;
         } else {
-            int i = state.getValue(POWER).intValue();
+            int i = state.getValue(POWER);
 
             if (i == 0) {
                 return 0;
@@ -327,7 +327,7 @@ public class BlockRedstoneWire extends Block {
     }
 
     public void randomDisplayTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
-        int i = state.getValue(POWER).intValue();
+        int i = state.getValue(POWER);
 
         if (i != 0) {
             double d0 = (double) pos.getX() + 0.5D + ((double) rand.nextFloat() - 0.5D) * 0.2D;
@@ -350,11 +350,11 @@ public class BlockRedstoneWire extends Block {
     }
 
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(POWER, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(POWER, meta);
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     protected BlockState createBlockState() {
