+++ src/main/java/net/minecraft/client/renderer/texture/TextureMap.java
@@ -219,7 +219,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                             Iterator iterator1 = list1.iterator();
 
                             while (iterator1.hasNext()) {
-                                int j4 = ((Integer) iterator1.next()).intValue();
+                                int j4 = (Integer) iterator1.next();
 
                                 if (j4 > 0 && j4 < abufferedimage.length - 1 && abufferedimage[j4] == null) {
                                     ResourceLocation resourcelocation = this.completeResourceLocation(resourcelocation1,
@@ -230,7 +230,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                                                 resourceManager.getResource(resourcelocation).getInputStream());
                                     } catch (IOException ioexception) {
                                         logger.error("Unable to load miplevel {} from: {}",
-                                                new Object[]{Integer.valueOf(j4), resourcelocation, ioexception});
+                                                new Object[]{j4, resourcelocation, ioexception});
                                     }
                                 }
                             }
@@ -257,10 +257,10 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
 
                     if (j3 < k) {
                         logger.warn("Texture {} with size {}x{} limits mip level from {} to {}",
-                                new Object[]{resourcelocation2, Integer.valueOf(textureatlassprite3.getIconWidth()),
-                                        Integer.valueOf(textureatlassprite3.getIconHeight()),
-                                        Integer.valueOf(MathHelper.calculateLogBaseTwo(k)),
-                                        Integer.valueOf(MathHelper.calculateLogBaseTwo(j3))});
+                                new Object[]{resourcelocation2, textureatlassprite3.getIconWidth(),
+                                        textureatlassprite3.getIconHeight(),
+                                        MathHelper.calculateLogBaseTwo(k),
+                                        MathHelper.calculateLogBaseTwo(j3)});
                         k = j3;
                     }
 
@@ -286,7 +286,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
 
             if (k2 < this.mipmapLevels) {
                 logger.warn("{}: dropping miplevel from {} to {}, because of minimum power of two: {}", new Object[]{
-                        this.basePath, Integer.valueOf(this.mipmapLevels), Integer.valueOf(k2), Integer.valueOf(j2)});
+                        this.basePath, this.mipmapLevels, k2, j2});
                 this.mipmapLevels = k2;
             }
 
@@ -315,7 +315,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                             return textureatlassprite1.getFrameCount() + " frames";
                         }
                     });
-                    crashreportcategory.addCrashSection("Mipmap levels", Integer.valueOf(this.mipmapLevels));
+                    crashreportcategory.addCrashSection("Mipmap levels", this.mipmapLevels);
                     throw new ReportedException(crashreport);
                 }
             }
@@ -330,8 +330,8 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                 throw stitcherexception;
             }
 
-            logger.info("Created: {}x{} {}-atlas", new Object[]{Integer.valueOf(stitcher.getCurrentWidth()),
-                    Integer.valueOf(stitcher.getCurrentHeight()), this.basePath});
+            logger.info("Created: {}x{} {}-atlas", new Object[]{stitcher.getCurrentWidth(),
+                    stitcher.getCurrentHeight(), this.basePath});
 
             if (Config.isShaders()) {
                 ShadersTex.allocateTextureMap(this.getGlTextureId(), this.mipmapLevels, stitcher.getCurrentWidth(),
@@ -451,7 +451,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                         this.basePath, location.getResourcePath(), ".png"))
                 : new ResourceLocation(location.getResourceDomain(),
                 String.format("%s/mipmaps/%s.%d%s", this.basePath,
-                        location.getResourcePath(), Integer.valueOf(p_147634_2_), ".png")));
+                        location.getResourcePath(), p_147634_2_, ".png")));
     }
 
     public TextureAtlasSprite getAtlasSprite(String iconName) {
@@ -715,11 +715,11 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                                 int i = dimension.width;
                                 int j = MathHelper.roundUpToPowerOfTwo(i);
 
-                                if (!map.containsKey(Integer.valueOf(j))) {
-                                    map.put(Integer.valueOf(j), Integer.valueOf(1));
+                                if (!map.containsKey(j)) {
+                                    map.put(j, 1);
                                 } else {
-                                    int k = ((Integer) map.get(Integer.valueOf(j))).intValue();
-                                    map.put(Integer.valueOf(j), Integer.valueOf(k + 1));
+                                    int k = (Integer) map.get(j);
+                                    map.put(j, k + 1);
                                 }
                             }
                         }
@@ -735,8 +735,8 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
         int l1;
 
         for (Iterator iterator = set1.iterator(); iterator.hasNext(); l += l1) {
-            int j1 = ((Integer) iterator.next()).intValue();
-            l1 = ((Integer) map.get(Integer.valueOf(j1))).intValue();
+            int j1 = (Integer) iterator.next();
+            l1 = (Integer) map.get(j1);
         }
 
         int i1 = 16;
@@ -745,8 +745,8 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
         Iterator iterator1 = set1.iterator();
 
         while (iterator1.hasNext()) {
-            int i2 = ((Integer) iterator1.next()).intValue();
-            int j2 = ((Integer) map.get(Integer.valueOf(i2))).intValue();
+            int i2 = (Integer) iterator1.next();
+            int j2 = (Integer) map.get(i2);
             k1 += j2;
 
             if (i2 > i1) {
