+++ src/main/java/net/minecraft/entity/monster/EntityEnderman.java
@@ -77,9 +77,9 @@ public class EntityEnderman extends EntityMob {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, Short.valueOf((short) 0));
-        this.dataWatcher.addObject(17, Byte.valueOf((byte) 0));
-        this.dataWatcher.addObject(18, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(16, (short) 0);
+        this.dataWatcher.addObject(17, (byte) 0);
+        this.dataWatcher.addObject(18, (byte) 0);
     }
 
     public void writeEntityToNBT(NBTTagCompound tagCompound) {
@@ -262,7 +262,7 @@ public class EntityEnderman extends EntityMob {
     }
 
     public void setHeldBlockState(IBlockState state) {
-        this.dataWatcher.updateObject(16, Short.valueOf((short) (Block.getStateId(state) & 65535)));
+        this.dataWatcher.updateObject(16, (short) (Block.getStateId(state) & 65535));
     }
 
     public boolean attackEntityFrom(DamageSource source, float amount) {
@@ -310,7 +310,7 @@ public class EntityEnderman extends EntityMob {
     }
 
     public void setScreaming(boolean screaming) {
-        this.dataWatcher.updateObject(18, Byte.valueOf((byte) (screaming ? 1 : 0)));
+        this.dataWatcher.updateObject(18, (byte) (screaming ? 1 : 0));
     }
 
     static class AIFindPlayer extends EntityAINearestAttackableTarget {
