+++ src/main/java/net/minecraft/village/Village.java
@@ -326,13 +326,13 @@ public class Village {
 
     public int getReputationForPlayer(String p_82684_1_) {
         Integer integer = this.playerReputation.get(p_82684_1_);
-        return integer != null ? integer.intValue() : 0;
+        return integer != null ? integer : 0;
     }
 
     public int setReputationForPlayer(String p_82688_1_, int p_82688_2_) {
         int i = this.getReputationForPlayer(p_82688_1_);
         int j = MathHelper.clamp_int(i + p_82688_2_, -30, 10);
-        this.playerReputation.put(p_82688_1_, Integer.valueOf(j));
+        this.playerReputation.put(p_82688_1_, j);
         return j;
     }
 
@@ -367,10 +367,10 @@ public class Village {
                 GameProfile gameprofile = playerprofilecache.getProfileByUUID(UUID.fromString(nbttagcompound1.getString("UUID")));
 
                 if (gameprofile != null) {
-                    this.playerReputation.put(gameprofile.getName(), Integer.valueOf(nbttagcompound1.getInteger("S")));
+                    this.playerReputation.put(gameprofile.getName(), nbttagcompound1.getInteger("S"));
                 }
             } else {
-                this.playerReputation.put(nbttagcompound1.getString("Name"), Integer.valueOf(nbttagcompound1.getInteger("S")));
+                this.playerReputation.put(nbttagcompound1.getString("Name"), nbttagcompound1.getInteger("S"));
             }
         }
     }
@@ -411,7 +411,7 @@ public class Village {
 
             if (gameprofile != null) {
                 nbttagcompound1.setString("UUID", gameprofile.getId().toString());
-                nbttagcompound1.setInteger("S", this.playerReputation.get(s).intValue());
+                nbttagcompound1.setInteger("S", this.playerReputation.get(s));
                 nbttaglist1.appendTag(nbttagcompound1);
             }
         }
