+++ src/main/java/net/minecraft/world/storage/MapStorage.java
@@ -119,7 +119,7 @@ public class MapStorage {
 
                     if (nbtbase instanceof NBTTagShort nbttagshort) {
                         short short1 = nbttagshort.getShort();
-                        this.idCounts.put(s, Short.valueOf(short1));
+                        this.idCounts.put(s, short1);
                     }
                 }
             }
@@ -132,15 +132,15 @@ public class MapStorage {
         Short oshort = this.idCounts.get(key);
 
         if (oshort == null) {
-            oshort = Short.valueOf((short) 0);
+            oshort = (short) 0;
         } else {
-            oshort = Short.valueOf((short) (oshort.shortValue() + 1));
+            oshort = (short) (oshort.shortValue() + 1);
         }
 
         this.idCounts.put(key, oshort);
 
         if (this.saveHandler == null) {
-            return oshort.shortValue();
+            return oshort;
         } else {
             try {
                 File file1 = this.saveHandler.getMapFileFromName("idcounts");
@@ -149,7 +149,7 @@ public class MapStorage {
                     NBTTagCompound nbttagcompound = new NBTTagCompound();
 
                     for (String s : this.idCounts.keySet()) {
-                        short short1 = this.idCounts.get(s).shortValue();
+                        short short1 = this.idCounts.get(s);
                         nbttagcompound.setShort(s, short1);
                     }
 
@@ -161,7 +161,7 @@ public class MapStorage {
                 exception.printStackTrace();
             }
 
-            return oshort.shortValue();
+            return oshort;
         }
     }
 }
